+++ RxSwift/Traits/PrimitiveSequence.swift
@@ -627,6 +627,20 @@ extension PrimitiveSequenceType where TraitType == MaybeTrait {
     }
 }
 
+extension PrimitiveSequence where Trait == CompletableTrait, Element == Never {
+    /**
+     Concatenates the second observable sequence to `self` upon successful termination of `self`.
+     
+     - seealso: [concat operator on reactivex.io](http://reactivex.io/documentation/operators/concat.html)
+     
+     - parameter second: Second observable sequence.
+     - returns: An observable sequence that contains the elements of `self`, followed by those of the second sequence.
+     */
+    public func concat(_ second: PrimitiveSequence<CompletableTrait, Never>) -> PrimitiveSequence<CompletableTrait, Never> {
+        return Completable.concat(self, second)
+    }
+}
+
 extension PrimitiveSequenceType where TraitType == CompletableTrait, ElementType == Never {
     /**
      Returns an empty observable sequence, using the specified scheduler to send out the single `Completed` message.
@@ -637,6 +651,44 @@ extension PrimitiveSequenceType where TraitType == CompletableTrait, ElementType
      */
     public static func empty() -> PrimitiveSequence<CompletableTrait, Never> {
         return PrimitiveSequence(raw: Observable.empty())
+    }
+    
+    /**
+     Concatenates all observable sequences in the given sequence, as long as the previous observable sequence terminated successfully.
+     
+     - seealso: [concat operator on reactivex.io](http://reactivex.io/documentation/operators/concat.html)
+     
+     - returns: An observable sequence that contains the elements of each given sequence, in sequential order.
+     */
+    public static func concat<S: Sequence>(_ sequence: S) -> PrimitiveSequence<CompletableTrait, Never>
+        where S.Iterator.Element == PrimitiveSequence<CompletableTrait, Never> {
+            let source = Observable.concat(sequence.lazy.map { $0.asObservable() })
+            return PrimitiveSequence<CompletableTrait, Never>(raw: source)
+    }
+    
+    /**
+     Concatenates all observable sequences in the given sequence, as long as the previous observable sequence terminated successfully.
+     
+     - seealso: [concat operator on reactivex.io](http://reactivex.io/documentation/operators/concat.html)
+     
+     - returns: An observable sequence that contains the elements of each given sequence, in sequential order.
+     */
+    public static func concat<C: Collection>(_ collection: C) -> PrimitiveSequence<CompletableTrait, Never>
+        where C.Iterator.Element == PrimitiveSequence<CompletableTrait, Never> {
+            let source = Observable.concat(collection.map { $0.asObservable() })
+            return PrimitiveSequence<CompletableTrait, Never>(raw: source)
+    }
+    
+    /**
+     Concatenates all observable sequences in the given sequence, as long as the previous observable sequence terminated successfully.
+     
+     - seealso: [concat operator on reactivex.io](http://reactivex.io/documentation/operators/concat.html)
+     
+     - returns: An observable sequence that contains the elements of each given sequence, in sequential order.
+     */
+    public static func concat(_ sources: PrimitiveSequence<CompletableTrait, Never> ...) -> PrimitiveSequence<CompletableTrait, Never> {
+        let source = Observable.concat(sources.map { $0.asObservable() })
+        return PrimitiveSequence<CompletableTrait, Never>(raw: source)
     }
     
     /**