+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ## Master
 
 * Adds `from` operator on "SharedSequence"
+* Adds `concat` operator on "Completable"
 * Adds `merge` operator on "Completable"
 * Adds `using` operator on "PrimitiveSequence"